Patch DMA0 to use a cluster for TCDs
DMA transfer control descriptors (TCD) may be represented as a
cluster. The commit adds a patch to all SVD files that defines
the clustering.

This ends up minimally transforming the output Rust. The changes
manifest as a renaming resembling

  TCD8_ATTR -> TCD_ATTR8

Howeverm the change may be useful for other users of these patched
SVDs. I know that svd2rust would take advantage of this change.