imxrt-ral check now checks all chip variant builds
The default feature flag for imxrt1062 was removed and imxrt-hal now requires
feature flags to enable each chip. Currently only support the imxrt106x series
as other chips had various missing registers that need to be dealt with. In
particular the imxrt101x does not have a PLL_ARM register.

Adds a make file for imxrt-hal to check various combinations of feature
flags as well.

The check targets can easily be expanded for other optional features
like rt, rtfm, and nosync if we want.

In addition splits .gitignore into each directory to avoid a mega
.gitignore.