Port iomux to imxrt-ral
Note that each pad requires unsafe modify_reg() calls to convert to a
particular pads alternate usage.

This should be safe because each pad has its own register pair (mux/pad) for
setting things up. So long as the only place where these registers are
used is in the iomuxc instance in struct Peripherals, this is like just
fine.