Add the I2C pins; remove Daisy traits
There are a couple of semi-related changes. A gpio's pad() method isn't
very useful, since there's no good way to work with the register
without using the macros. The commit removes a gpio's pad() method.
We'll need to replace the call to pad() with direct ral::write_reg!
macros. This commit demonstrates how that might work for other
peripherals.

Remove the daisy module. There's no need for this independent trait;
the UART module shows that it's easier to just put the daisy
configurations directly into the Pin implementation. It's just over-
head and noise for the developer. We will need to solve this in other
peripherals as we move forward (practially, this will only apply to
the SPI peripheral, once ported. I'll take care of that in the port).

This commit should be reviewed for correctness of the I2C pin
configuration.