Remove PAD register accessors from GPIOs
See the previous commit for more justification. There's no safe
way for us to access and modify a GPIO's PAD_CTL register without
using the RAL macros. The commit removes those register from the
pad! macros, and it removes access of the register from the GPIO.
Developers are expected to directly modify the PAD_CTL register on
a pin-by-pin basis. The I2C port in the previous commit demonstrates
what that might look like.

This commit may be reverted if there's another way of doing this!