Port I2C implementation to the RAL API
The commit updates the I2C master implementation to use the RAL API
for register access. We remove the PAC calls and add corresponding
calls to the RAL crate. When reviewing this commit, we should see
a 1:1 correspondence from the PAC call to the RAL call, unless I noted
otherwise.

As part of this commit, I removed some of the CCM type overhead;
notably, the Ticks type. Since the RAL favors symbols of u32 types,
it's harder to justify having our own stronger types, particularly
for things that aren't exposed from our HAL. The Ticks type was one
of those types. The I2C clock speed configuration routine is changed
to use u32s, rather than the previously-specified stronger types.