Port PIT timers
The PIT timers needed some internal changes to support the lack of
register clustering in the RAL. We extend the existing Channel trait
to provide implementations of the register manipulations that we need.
We then call onto those trait methods in the PIT struct. We should
check the port for consistency.