Port the PWM to the RAL
This peripheral was the most complicated to port. The previous
implementation was using the fact that svd2rust would represent
SVD `<cluster>`s as structs. The RAL just provides registers and
different symbols. We needed to account for this implementation change
in the port.

The new approach should also be safer. Tom identified some unsafety
that I wasn't considering in my first implementation. The new approach
should fix the safety issues in the PWM peripheral.

Most PWM operations require access to both submodule- and peripheral-
level registers. To enforce this, we break-out a PWM Handle, and
PWM submodules. Operations that require global PWM registers require
mutable access to the Handle. To let people wrap their Handles in a
Mutex<RefCell<Handle<module::_1>>>, we generalize some APIs by using
DerefMut, which should let us take anything that owns a Handle and
provides exclusivity. With this approach, we shouldn't need to
introduce our own critical sections in our implementation.

We implement embedded_hal::Pwm, rather than embedded_hal::PwmPin. It
was easier to do than the PwmPin implementation. Since most of the
interesting methods require access to both peripheral-level and
submodule-level registers, we introduce a `Controller`. The
`Controller` implements the Pwm trait, and it wraps a mutable Pins
and a type that can deref to a mutable Handle. A user can only control
PWM pin settings with the `Controller` type. Once a user is done setting
register values, she may drop the `Controller`.

Tested this commit on hardware!