Write periodic clock divider to CSCMR1
I was finishing the rounds of hardware testing on the Teensy 4, and I
found that timings based on PITs were wrong. I tracked it down to the
CCM module, and how we modify the PIT peripheral's clock inputs. We
weren't setting the divider. The commit updates the modify_reg! call
to set the divider as part of the periodic clock configuration flow.