Add general purpose timer
The general purpose timers (GPT) complement the PIT timers. The timers
simply count up, then wrap around. They can be used to measure
duration, or interrupt on intervals.

There are two GPTs. Each has three comparison registers, called
'output compare register.' When the running counter matches a value
in an OCR, the processor can generate an interrupt. It may also
generate an output signal on a pin; although, this isn't implemented.

The GPTs can also capture the counter on an *input* signal. Each GPT
accepts two GPIO inputs. The input capture can generate an interrupt.
This is still TODO.

Tested by an example in the teensy4-rs project. The test showed that
we can interrupt, and set the compare duration.

The GPTs and the PITs share the same root clock. To support clocking
of both peripherals, I needed to make a breaking change to the PIT's
`clock()` method. The method now accepts a mutable reference, rather
than a value, to the handle that indicates proper periodic clock
configuration. Crate implementers are responsible for calling the
correct method to enable the clock gates for the timer instance.