Do not change default wait mode behavior
Let the users decide the behavior of the GPTs. No need to override
the default behavior. We documented the default behavior, then say
how a user might want to change that.

We're still overriding the free-run mode, since the API doesn't let
a user account for the 'special' channel 1. Might want to rethink
this later...