Cache the entire RAL directory for CI builds
CI errors indicate that automated builds are failing because the RAL's
Cargo.toml isn't there. We removed the Cargo.toml file in a previous
commit, since we opted to exclude all auto-generated RAL files from
git. CI builds was probably relying on that file being in git, and
now the build fails.

By including the entire RAL directory in the cache, it will capture
not only the auto-generated Rust code, but also the RAL's Cargo.toml.