Fix docs.rs documentation generation
The imxrt-hal docs aren't being published on docs.rs! See

  https://docs.rs/crate/imxrt-hal/0.2.1/builds

The logs indicate it's because we didnt' specify a feature-flag
to be used when generating the docs. The commit adds the feature
that docs.rs should use when generating the crate documentation.

This will make its way into docs.rs the next time we make a HAL
release.