Decouple transfer iteration setting from buffers
We may want to set the number of transfer iterations independent of
the source / destination buffer size. This lets a higher-level DMA
abstraction specify the number of transfer iterations independent
of the buffer size. This could be used for a hypothetical peripheral-
to-peripheral transfer, which would need to specify how many elements
to transfer across the peripherals.