Fix obvious issues in memcpy
- should set disable on completion, since we need to explicitly
  disable the channel, which we do as of this commit
- move start() call before fence. Unsure if this has an impact at
  the moment, but the reference manual example seemed to group all
  TCD operations next to each other
- call complete on the provided buffers

At this commit, `Memcpy` doesn't work. Or, it didn't work in my
testing on a Teensy 4. I modified the `Memcpy` implementation to
call `start()` on the DMA channel in a busy-loop on `is_complete()`,
and that has the effect of driving the transfer to completion. My
current thinking is that we need to use a single major loop for in-
memory DMA transfers, similar to what's documented in the reference
manual. There's no external peripheral to request DMA services, so
we need to move all the data in one call to `start()`.