Add proper support for DMA memcopy
This spreads the channel setup throughout the peripheral and buffer
modules. We define a `Transfer` struct, which names and documents
the values that go into defining a transfer. We separate minor loop
iterations from major loop iterations, which let us support a memcpy
with a single DMA service request.

The memcpy example I tested copied memory across two circular
buffers. I found that I was both incorrectly specifying the modulo
value, and not correctly aligning my buffer. Both modulo, and the
alignment, need to account for the size of the element! The former
is fixed here; the latter can be enforced by an error at run-time.