Correctly check for circular buffer alignment
The runtime check wasn't accounting for the element size. It now
checks for the element size. Updated the docs.

We can't reliably run the mis-alignment example, since there's a
chance that the linker could place the buffer at a memory address
that satisfies the alignment. The test could sporadically fail. We
instead check that the example compiles.

Added docs that describe the variability of this runtime-check
approach.