Use DMA multiplexer for DMA memory copies
The DAM multiplexer can treat a DMA channel as "always on." When a
channel is always on, it doesn't need explicit re-activation, or
service requests, to drive the major loop to zero. This means we can
accomplish a memory copy across multiple major loops, not just one,
large major loop, without needed explicit software re-activation.

At this commit, all DMA transfers run through the DMA multiplexer.
We have a consistent method for disabling transfers; we disable
the 'enable' flag in ERQ. This undoes the changes to the enable
method introduced in the previous commit.