Document how many DMA channels are initialized
We only initialize the first CHANNEL_COUNT channels. If the code runs
on a 1011, that means only the lower half of the array will have
valid channels. The rest are none.

I couldn't think of an elegant way to signal this at compile time.
Users are free to use the CHANNEL_COUNT constant to do something
fancies. Or, we could consider signaling a compile-time configuration
from a build script.