iomuxc: fix feature naming discrepancies
The "x" suffix is inconsistent with the NXP product naming scheme.
Since it diverges, it's not clear what chips are supported for the
"imxrt101x" feature.

The commit updates the imxrt-iomuxc crate to use product naming
conventions. The feature flag migration follows

- "imxrt106x" => "imxrt1060"
- "imxrt101x" => "imxrt1010"

This change is reflected in the filesystem, and in all documentation
and code updates.