Refactor the GPIO module
The commit revises the GPIO API to require the RAL GPIO instance
before permitting pin construction. By not requiring the RAL GPIO
instance, we could allow users to both allocate pins, and also modify
GPIO memory through the instance. This approach should require that
the user first give us the RAL instance, so that we have the only
instance in the program. We then use that GPIO register instance
to further allocate GPIO pins.