Add CCM and DCDC APIs
The CCM API is broken into a few separate modules:

- clock_gate provides a named API for interfacing peripheral clock
  gates.
- clock_tree builds on root clock modules, providing presets for system
  and peripheral clocks.
- all other modules are root clock modules.

ccm doesn't try to encapsulate the CCM and CCM_ANALOG peripherals;
instead, it just works with references to these peripherals.

The DCDC follows a similar design, though it's very simple today.

The clock_tree API depends on the RunMode configuration.