Add preliminary support for 1170 MCUs
The hal_led example works on the CM7 of an 1170 EVK. This commit
includes some code movement to separate the 10xx code from the 11xx
code, both in the HAL and the board.

The DMA pointers that were hard-coded in the DMA module were wrong for
the 1170. This commit changes how the common DMA code references those
pointers. It grabs them from the RAL, and casts them. This should have
been the default :P

We'll likely need true clocking support before we can bring up other
1170 peripherals. That's coming next.