Move DMA peripheral impls into chip module
The DMAMUX signals change when moving from 10xx to 11xx, so we'll need a
chip feature before compiling the DMA peripheral impls. Move the
implementations into the chip DMA module, and select the set of DMAMUX
signals at compile time.

As of this commit, the 1170EVK can correctly run the RTIC DMA UART
example.