Explore type-erased LPUART
AnyLpuart removes instance numbers and pins from the Lpuart type
signature. These types can be easier to work with in some contexts,
like when you want to accept any LPUART and (safely) store it in a
static.

Once you have an AnyLpuart, you cannot safely re-acquire the resources.
We drop the zero-sized pin objects, and we don't replace them with their
erased-but-sized counterparts. We also lose the N const, so we can't
safely re-create the instance (there's patterns to work around this;
they can be implemented externally before they're integrated here).
Keeping the AnyLpuart to the size of a reference lets us have
null-pointer optimizations.

Strongly-typed Lpuart objects deref to AnyLpuart, so most of the
implementation now resides on AnyLpuart. To make embedded-hal
compatibility easier, we implement the traits on both AnyLpuart and
Lpuart<P, N>. Without this, users would need to manually deref to
AnyLpuart, and that would be annoying.

I'm mulling over this pattern, and I plan to extend something similar to
the other peripherals.