Add async DMA examples for LPUART, LPSPI
Move the helper functions out of the dma module and into board. These
were helpers from when the dma module was its own crate. Updated the
doc examples with async blocks to demonstrate await syntax.

Add a new DMA channel for all boards so that we can demonstrate
full-duplex SPI.