Remove cirular BBBuffer from imxrt-log
The LPUART backend now uses the BBBuffer as linear memory, not as a
circular buffer, for DMA transfers. This removes the dependency on my
bbqueue fork which contained a small library change to support circular
buffers. We can also remove the customized Align adapter for aligning
that buffer, along with some of the initialization behavior.