Add button GPIO input to 1010EVK, Teensy 4
Two new examples demonstrate how to use a GPIO as an input. The board
module configures the pad with a PU / PD, and the examples toggle an LED
state based on the input's state or change.

The 1010EVK has the easiest experience, since it already has an on-board
user button. I'm choosing Pin 7 of the Teensy 4 for a external switch.

The two boards now have a BOARD_BUTTON interrupt handler. I'm not sure
how this will work if we start adding more GPIO inputs. The approach I'm
taking bets that this isn't going to happen. Otherwise, we would need to
figure how how to keep the examples consistent while varying how ISRs
are mapped to their vectors.

There's no support for the 1170EVK in this commit. The buttons on that
board are routed to IOMUXC_SNVS pads, which we're not exposing from
imxrt-iomuxc. At this time, it's non-trivial to introduce those pads, so
the next commit adds a HAL work-around to avoid the limitation.