Add LPUART driver
Peripheral instance and pins are available in the type system. Use the
without_pins() constructor to allocate a driver without pins. Similar to
the GPIO driver, this is helpful when imxrt-iomuxc doesn't yet support
your device, or when you want more flexibility in your types.

As with other new drivers, I'm trying to use shared references when
actions on the peripheral are safe across execution contexts. For
example, reading and writing bytes from the FIFO can occur with shared
references; but clearing the FIFO needs an exclusive reference. I'm
curious to see how this works out in practice. Here's some
considerations:

- Users can distribute shared references, and those shared references
  can be more useful.
- Users can elect to take exclusive references even when all the methods
  they use could work with shared references. Users should do this to
  ensure that state changes on their drivers are consistent; we're not
  always doing that for them.
- Maintainers need to be extra careful; we're favoring shared references
  more than exclusive references, and we need to make sure that's
  correct.