Add chip-specific DMA implementations
We don't know either of

- the DMAMUX signals for a peripheral
- the channel count

until the user selects a chip family. Based on the chip family, this
module implements DMA support for peripherals, and exposes the DMA
channel API to the user.

I'm generally not sure if these DMA future operations are proper
futures. Specifically, there's a state change in the DMA channel and
peripherals during the future construction, and before the first poll of
the future. If my understanding of futures is correct, they're expected
to be inert until the first poll call, and construction of a future
shouldn't have side effects. The actual DMA transfer isn't happening, so
these state changes represent the driver / DMA channel configurations.
It's something to revisit later.

Lane provided the original ADC DMA support.

Co-authored-by: Lane Kolbly <lane@rscheme.org>