Add CCM APIs for 10xx chips
The CCM API is a small step above the imxrt-ral API. It exposes free
functions and enums to configure clock selections and dividers. The
clock_gate module gives users a way to manipulate clock gates. I've
explored other ideas for a CCM driver, but I didn't like any of them as
much as this lower-level API.

There's a clock_tree module that implements a root clock selection
policy based on the user-specified run mode. This was helpful for
prototyping, but I'm not fully committed to its maintenance and
extension to the 11xx MCU. I'm keeping it around to see if others find
it valuable.