Add basic CCM driver for 11xx chips
The CCM on the 11xx is much more complicated than the 10xx chips. This
basic driver assumes that you're only using unassigned modes for all
clocks (the default out of reset). The clock_gate and clock_tree APIs
try to mimic the APIs in the 10xx modules.

One notable omission: there's no support for turning on PLLs. So clocks
are using whatever is available out of reset. As with the 10xx
clock_tree API, I'm not fully sold on having these clock_tree policies
in the HAL. They're provided for consistency with the 10xx API.

The (maybe) redeeming quality of the clock_gate API is that (I think)
the locator offsets are the same no matter the clock mode. This could
provide a simple way to support different clock gating behaviors for
clock modes.