Add imxrt-hal package root, build script
The package root tries to re-export modules as-is. If a driver module is
comprised of common and chip-specific code, we attach the module
documentation to a named module that re-exports the items. This means
that drive module contents may vary by chip features.

To simplify some conditional compilation, the build script emits
'family' configurations. They're equivalent to doing an 'any' cfg, but
easier to read.