Move RunMode, clock_tree, power APIs to board
These APIs were helpful for prototyping, but I'm not fully convinced
that they belong in imxrt-hal. I'm concerned the HAL won't be able to
support all users, and that it's better for boards / users to define
their clock tree settings themselves, using the ccm APIs.

The boards were using this API, so this commit moves the APIs to the
boaard module. If users find value in these APIs, we can always move
these back into imxrt-hal and release as a non-breaking change.