Fix spelling of "trace clock" for 1010 CLKO2
We can't add a new variant to the enum, since the discriminates would be
the same value. But, the new spelling can exist with the old spelling if
we use an associated constant. This lets us introduce the change in a
backwards-compatible manner. The deprecation notice tells users to fix
the spelling before a future removal.

Tested with a new, small integration test. Notice how it's still using
the incorrect spelling.