Add USB serial loopback example for board testing
It uses usbd_serial::SerialPort, the higher-level USB class provided by
the package. Using this with a high-speed USB device has some caveats;
see the example documentation.

I'm routing debug logging to the board's UART console. Internally,
imxrt-usbd uses the log package for debugging, so this means we can
access those logs. Ideally, we would use defmt to reduce the logging
overhead in the driver; that's a future TODO.

Tested on a Teensy 4, 1010EVK, and 1170EVK (CM7). Only showed that UART
logging works on the 1010EVK.