Increase stack sizes for most boards
In a debug build, the rtic_usb_test_class example blows up just over
16KiB of stack in the `init` procedure. That's all the stack we're
allocating (by default). This commit makes it easier to execute debug
builds. This commit excludes the 1010EVK; this would take up one of the
four FlexRAM banks, and debug builds are already crammed on that MCU.