Remove PCS0 from Lpspi::new pin requirements
When we support the embedded-hal 1.0 traits, we should distinguish
between the bus -- having the data and clock pins -- and the device(s),
which have the chip select. Removing the chip select type state is the
first step in supporting this interface.

Temporarily enable PCS0 in the board setup routines. This keeps the
examples working as expected.