Ensure LPSPI exchanges fill the receive buffer
`exchange` would the transmit buffer and break the loop without reading
anything. We could demonstrate the issue with the `hal_spi` example in
this repo. This commit corrects the loop condition so that it waits for
all receive data.

Terminate the continuous transaction once we enqueue the final command.
This works around a hardware defect that would prevent us from reading
the final received word during continuous transfers. Now that we're
terminating within the loop, we can remove the terminating command at
the end of the method.

Tested on a Teensy 4.1 by the author. The committer also tested on an
1170EVK with the `hal_spi` loopback example. Example works as expected
when SDI and SDO are connected. When they're disconnected, the MCU
signals "data mismatch" over the serial console.