Increase imxrt-log default bulk MPS to 512
Community troubleshooting on Windows revealed that the USB device must
use a bulk endpoint max packet size (MPS) of 512 bytes if the device is
high speed. Since imxrt-log's USB back-end is already high-speed by
default, this commit increases the default MPS to 512 bytes.

Tested using the rtic_logging and hal_logging examples on a Teensy 4.
I've not tested against a Windows host, which is where this issue
manifested.