Use defmt-rtt and panic-probe for evk boards
The evk boards have a built in debug probe. When running firmware with probe-rs
defmt-rtt provides a very fast channel of low cost debug println's that everyone
finds really helpful.

Using panic-probe with defmt-rtt has a second benefit in allowing for programs to
inform probe-rs the firmware has exited cleanly. This can be used for writing samples
and tests that exit.

This also makes the imxrt-log defmt feature optional to the boards otherwise
the two implementations of the defmt logger cause duplicate symbols to show up.

Tested on an 1010evk and 1060evk.