Detect imxrt-log's defmt feature in board builds
We can't detect a dependency's features in our build. We can show this
by modifying the rtic_logging example to use the Frontend::Defmt
frontend, then building the example for a Teensy 4. The example fails to
build.

We can only detect our own features in our build. Add a board feature
that we can enable for a board, and rewrite the conditional includes to
use the feature.