Expose ENET through 1170EVK
It's the ENET instance that does not support 1Gbit operation. The PHY
reset and configuration is hard-coded into the board module, and it's
fixing the PHY for 10Mbit/s operation. If I go faster, the receive
pathway doesn't work (but the transmit pathway seems fine). One thing
I'd like to rule out is the clock quality; bringing up the ENET PLL is a
TODO.