Enable PLL1 with basic setpoint support
Enable the PLL_LDO in all setpoints, and enable PLL1 in setpoint 1.
Since I'm only using setpoints for PLL_LDO and PLL1 bring-up, I'm
implementing the rest of the GPC state machine in software (specifically
for the ENET clocks). This seems sufficient for enabling PLL1 and
getting a clean 50MHz ENET reference clock. I can increase the ENET
transfer speeds in the MAC and PHY.

The GPC, PMU, and CCM APIs are a bare-minimal. They should be introduced
in a separate commit / PR. But they're good enough for prototyping.