Address latest clippy warnings
Clippy warns when we're re-exporting an empty module, since it detects
our import of that empty module is "unused." I'm not a fan of leaking
the family selection into the top-level module, but I also want signal
if something is exported through the chip module and never used, like a
pub(crate) item. This is the easiest fix to read and write.

The alternates were to hide this complexity down in the `chip/none.rs`
module. I tried a doc(hidden) type that would be re-exported through all
the other modules, then marked the whole none module as allowing unused
imports. Looked strange and took more lines than what's shown here, but
it did solve the problem.

Another approach: change the module structures. We make all chips
responsible for re-exporting common. The none chip would only re-export
common, so there's no more unused imports, and no more empty modules for
the none build. This is a bigger refactor for another time.