Add weekly CI linting job
I've carried this in my fork for a little while, using it as a way to
get ahead of clippy issues. But I don't keep my fork's main up to date,
so I might miss things. I'll keep an eye on it upstream, and other folks
can help out too.