Update examples to RTIC v2
The new framework has support for async tasks. It might be helpful for
demonstrating async drivers, or for users who are comparing RTIC's
latest documentation against our examples.

This is a minimal migration, and it's not trying to change any example's
behavior. Let me know if you notice something different.

Tested on a 1010EVK using the `_pit`, `_defmt_rtt`, and `_logging`
examples.  Also tested `_rtic_controller` on an 1170EVK, and
`_rtic_device` on a 1010EVK.