Test continuous xfer-then-read behavior
When we change the transmit / receive masks within a continuous
transaction, we might demonstrate an LPSPI hardware defect. In the
example shown here, the LPSPI SCK stalls when transmitting the final bit
of the final byte, preventing the read from executing.

Right now, I'm not sure if this is the LPSPI continuous transaction
hardware defect, or if there's an issue in this new driver. I'll shoot
for a deeper review later.