Remove static muts from logging package
Upcoming clippy versions have more agressive warnings about static mut,
so I'm removing them from the logging package. The defmt frontend uses a
custom static with interior mutability. The implementation is otherwise
the same.

I refactored the rest of the package to use `&'static mut`s produced by
the static_cell package. This lets us remove the explicit uninitialized
memory and the mutable statics. The details are unchanged.

Tested the rtic_logging example on an 1170EVK: logging frontend with
both backends, and the defmt frontend with just the LPUART frontend.