Fix LPSPI word unpacking in receive operations
LPSPI receive operations are reversing byte and half-word elements when
they place data in the caller's buffer. We can show this using the
revised loopback tests on an older LPSPI driver.

The unpacking routines need to consider the frame boundary and the bit
order when they deconstruct the 32 bit LPSPI word.