Don't assume LPSPI FIFO size
This specific configuration isn't different on the 1180. However, the
LPI2C and LPUART FIFOs are 2x and 4x larger, respectively, on the 1180.
I was afraid that I was using constants in the LPI2C and LPUART drivers,
like I was here. But those two drivers look OK; it was just LPSPI that's
assuming a FIFO size.